Paypal

+++ api/src/controllers/Comment.js
@@ -1,44 +1,31 @@
 const { Post, User, Comment } = require("../db");
+const { addNotification } = require("./Inboxes");
 
-// const getComment = (req, res, next) => {
-//     const id = req.params.id;
-
-//          Comment.findAll({
-//             include: [
-//                 {
-//                 model: Post,
-//                 attributes: ["title", "id"],
-//                 // through: {
-//                 //     attributes: []
-//                 // }
-//                 },
-//                 {
-//                 model: User,
-//                 attributes: ["first_name", "last_name", "id"],
-//                 }
-//             ]
-//         }).then(comment => {
-//             if(id){
-//                 let commentId = comment.filter(el => el.id == id);
-//                 commentId.length ? res.status(200).send(commentId) : res.status(400).send("Comment not found")
-//             }
-
-//            return  res.send(comment)})
-//         .catch(error => next(error))
-
-// }
-
-
-const addComment = async(req, res, next) => {
-    const { idPost, idUser } = req.params
-    // const idUser = req.idUser
-    const obj = {};
+const addComment = async (req, res, next) => {
+  const { idPost, idUser } = req.params;
+  const { message } = req.body;
+  const obj = {};
 
   try {
-    const createdInPost = await Post.findByPk(idPost, { include: [User] });
+    const createdInPost = await Post.findByPk(idPost, {
+      include: [User, Comment],
+    });
     const createdBy = await User.findByPk(idUser);
+    if (!createdInPost || !createdBy)
+      return res.status(404).send("Parametros invalidos");
 
-    const newComment = await Comment.create(req.body);
+    let exist = createdInPost.dataValues.comments.find(
+      (e) => e.dataValues.message === message
+    );
+    if (exist) {
+      return res
+        .status(400)
+        .send("Rechazado, ya existe este comentario en el posteo");
+    }
+
+    const newComment = await Comment.create({
+      message,
+    });
     createdBy.addComment(newComment);
     createdInPost.addComment(newComment);
 
@@ -47,9 +34,19 @@ const addComment = async(req, res, next) => {
     obj.rating = newComment.dataValues.rating;
     obj.user = {
       first_name: createdBy.dataValues.first_name,
-      id: createdBy.dataValues.id,
       last_name: createdBy.dataValues.last_name,
+      id: createdBy.dataValues.id,
     };
+    obj.notification = false;
+
+    // if (createdInPost.dataValues.user.id !== createdBy.dataValues.id) {
+      addNotification(
+        "comment",
+        newComment.dataValues.id,
+        createdInPost.dataValues.user.id
+      );
+      obj.notification = true;
+    // }
 
     res.send(obj);
   } catch (error) {
@@ -57,9 +54,22 @@ const addComment = async(req, res, next) => {
   }
 };
 
-const updateComment = (req, res, next) => {
-  const { idComment } = req.params;
+const updateComment = async (req, res, next) => {
+  const { idComment, idUser } = req.params;
   const { message, rating } = req.body;
+  const comment = await Comment.findByPk(idComment, { include: [User] });
+  const ownUser = await User.findByPk(idUser);
+
+  if (!comment || !ownUser) return res.status(404).send("Datos no encontrados");
+
+  if (comment.dataValues.user.id !== ownUser.id) {
+    return res
+      .status(400)
+      .send(
+        "Accion denegada, solo el propietario puede actualizar el comentario"
+      );
+  }
+
   return Comment.update(
     { message, rating },
     {
@@ -75,20 +85,35 @@ const updateComment = (req, res, next) => {
     .catch((error) => next(error));
 };
 
-const deleteComment = (req, res, next) => {
-  const { idComment } = req.params;
-  return Comment.destroy({
-    where: {
-      id: idComment,
-    },
-  })
-    .then(() => res.status(200).send("Comment deleted successfully"))
-    .catch((error) => next(error));
+const deleteComment = async (req, res, next) => {
+  const { idComment, idUser } = req.params;
+  try {
+    const comment = await Comment.findByPk(idComment, { include: [User] });
+    const user = await User.findByPk(idUser);
+
+    if (!comment || !user) return res.status(404).send("Datos no encontrados");
+
+    if (comment.dataValues.user.id === user.id || user.isAdmin) {
+      await Comment.destroy({
+        where: {
+          id: idComment,
+        },
+      });
+    } else
+      return res
+        .status(400)
+        .send(
+          "Accion denegada, solo el propietario puede eliminar el comentario"
+        );
+
+    res.send("Comment deleted successfully");
+  } catch (error) {
+    next(error);
+  }
 };
 
 module.exports = {
   updateComment,
   deleteComment,
   addComment,
-  // getComment,
 };
