+++ api/src/controllers/Post.js
@@ -39,8 +39,9 @@ const getPost = (req, res, next) => {
     .then((post) => {
       if (idPost) {
         let postId = post.filter((el) => el.id == idPost);
+
         return postId.length
-          ? res.send(postId)
+          ? res.send(postId[0])
           : res.status(404).send("question not found");
       }
       if (title) {
@@ -57,15 +58,26 @@ const getPost = (req, res, next) => {
 };
 
 const addPost = async (req, res, next) => {
+  const { idUser } = req.params;
+  const { title, message, tag, module } = req.body;
   try {
-    const { idUser } = req.params;
     const createdBy = await User.findByPk(idUser);
 
-    const { title, message, rating, tag, module } = req.body;
+    const exist = await Post.findAll({
+      where: {
+        title,
+        message,
+      },
+    });
+
+    if (exist.length)
+      return res
+        .status(400)
+        .send("Rechazado, ya existe un posteo con el mismo titulo y mensaje");
+
     const postCreated = await Post.create({
       title,
       message,
-      rating,
     });
     const tags = await Tag.findAll({
       where: {
@@ -80,26 +92,49 @@ const addPost = async (req, res, next) => {
     postCreated.addTag(tags);
     section[0].addPost(postCreated);
     createdBy.addPost(postCreated);
-    return res.send("Post done successfully");
+    res.json(postCreated);
   } catch (error) {
     next(error);
   }
 };
 
 const updatePost = async (req, res, next) => {
-  const { idPost } = req.params;
-  const { title, message, rating, tag } = req.body;
+  const { idPost, idUser } = req.params;
+  const { title, message, tag } = req.body;
   try {
-    const postUpdate = await Post.findByPk(idPost);
-    const allTags = await Tag.findAll({
-      where: {
-        name: tag,
-      },
-    });
-    console.log(allTags);
-    await postUpdate.setTags(allTags);
+    const postUpdate = await Post.findByPk(idPost, { include: [User, Module] });
+    const ownUser = await User.findByPk(idUser);
+
+    if (!postUpdate || !ownUser)
+      return res.status(404).send("Datos no encontrados");
+
+    if (ownUser.id !== postUpdate.dataValues.user.id) {
+      return res
+        .status(400)
+        .send("Accion denegada, solo el propietario puede actualizar el post");
+    }
+
+    if (tag) {
+      const allTags = await Tag.findAll({
+        where: {
+          name: tag,
+        },
+        include: [Module],
+      });
+      let pass = allTags.find(
+        (e) => e.dataValues.module.name !== postUpdate.dataValues.module.name
+      );
+      if (Boolean(pass))
+        return res
+          .status(400)
+          .send(
+            `Tags deben pertenecer a Modulo ${postUpdate.dataValues.module.name}`
+          );
+      await postUpdate.setTags(allTags);
+    }
+
     const updateSuccess = await Post.update(
-      { title, message, rating },
+      { title, message },
       {
         where: { id: idPost },
         raw: true,
@@ -113,15 +148,63 @@ const updatePost = async (req, res, next) => {
   }
 };
 
-const deletePost = (req, res, next) => {
-  const { idPost } = req.params;
-  return Post.destroy({
-    where: {
-      id: idPost,
-    },
-  })
-    .then(() => res.status(200).send("post deleted successfully"))
-    .catch((error) => next(error));
+
+const deletePost = async (req, res, next) => {
+  const { idPost, idUser } = req.params;
+  try {
+    const post = await Post.findByPk(idPost, { include: [User] });
+    const user = await User.findByPk(idUser);
+
+    if (!post || !user) return res.status(404).send("Datos no encontrados");
+
+    if (post.dataValues.user.id === user.id || user.isAdmin) {
+      await Post.destroy({
+        where: {
+          id: idPost,
+        },
+      });
+    } else
+      return res
+        .status(400)
+        .send("Accion denegada, solo el propietario puede eliminar el posteo");
+
+    res.send("Post deleted successfully");
+  } catch (error) {
+    next(error);
+  }
+};
+
+
+const finishedPost = async (req, res, next) => {
+
+
+  try {
+
+    const { idPost } = req.params;
+    const { finished } = req.body;
+
+    const post = await Post.findByPk(idPost, { include: [User] });
+    console.log('hola', finished)
+
+    if (!post)
+      return res.status(404).send("Datos no encontrados");
+
+    const postFinished = await Post.update(
+      { finished },
+      {
+        where: { id: idPost },
+        // raw: true,
+      }
+    );
+
+    console.log('Este es el post finalizado:', postFinished);
+
+    return postFinished[0] === 1
+      ? res.send("Post actualizado con exito!")
+      : res.status(400).send("No se pudo actualizar el Post");
+  } catch (error) {
+    console.log(error)
+  }
 };
 
 module.exports = {
@@ -129,4 +212,5 @@ module.exports = {
   addPost,
   updatePost,
   deletePost,
+  finishedPost
 };
